extract column and schema_creation from mysql2_adapter
