+++ activerecord/test/cases/column_definition_test.rb
@@ -39,34 +39,34 @@ module ActiveRecord
 
       if current_adapter?(:Mysql2Adapter)
         def test_should_set_default_for_mysql_binary_data_types
-          binary_column = Mysql2Adapter::Column.new("title", "a", Type::Binary.new, "binary(1)")
+          binary_column = Mysql2AdapterColumn.new('title', 'a', Type::Binary.new, 'binary(1)')
           assert_equal "a", binary_column.default
 
-          varbinary_column = Mysql2Adapter::Column.new("title", "a", Type::Binary.new, "varbinary(1)")
+          varbinary_column = Mysql2AdapterColumn.new('title', 'a', Type::Binary.new, 'varbinary(1)')
           assert_equal "a", varbinary_column.default
         end
 
         def test_should_not_set_default_for_blob_and_text_data_types
           assert_raise ArgumentError do
-            Mysql2Adapter::Column.new("title", "a", Type::Binary.new, "blob")
+            Mysql2AdapterColumn.new('title', 'a', Type::Binary.new, 'blob')
           end
 
           assert_raise ArgumentError do
-            Mysql2Adapter::Column.new("title", "Hello", Type::Text.new)
+            Mysql2AdapterColumn.new('title', 'Hello', Type::Text.new)
           end
 
-          text_column = Mysql2Adapter::Column.new("title", nil, Type::Text.new)
+          text_column = Mysql2AdapterColumn.new('title', nil, Type::Text.new)
           assert_equal nil, text_column.default
 
-          not_null_text_column = Mysql2Adapter::Column.new("title", nil, Type::Text.new, "text", false)
+          not_null_text_column = Mysql2AdapterColumn.new('title', nil, Type::Text.new, 'text', false)
           assert_equal "", not_null_text_column.default
         end
 
         def test_has_default_should_return_false_for_blob_and_text_data_types
-          blob_column = Mysql2Adapter::Column.new("title", nil, Type::Binary.new, "blob")
+          blob_column = Mysql2AdapterColumn.new('title', nil, Type::Binary.new, 'blob')
           assert !blob_column.has_default?
 
-          text_column = Mysql2Adapter::Column.new("title", nil, Type::Text.new)
+          text_column = Mysql2AdapterColumn.new('title', nil, Type::Text.new)
           assert !text_column.has_default?
         end
       end
